file_structure: refactored lobby direcotry structure
