+++ backend/src/lobby/lobby.gateway.ts
@@ -8,11 +8,15 @@ import {
 } from "@nestjs/websockets";
 import { ClientEvents, ServerEvents } from "./events/lobby.events";
 import { LobbyService } from "./lobby.service";
-import { AuthenticatedSocket, ServerPayloads } from "./types/server.type";
+import { AuthenticatedSocket, ServerPayloads } from "./types/lobby.type";
 import { Server } from "socket.io";
-import { LobbyValidationPipe } from "./lobby.creator";
-import {LobbyDto} from "./dtos/lobby.dto";
+import { LobbyValidationPipe } from "./dto/lobby.pipe";
+import { JoinLobbyDto, LeaveLobbyDto, LobbyDto } from "./dto/lobby.dto";
+import { ValidationPipe } from "@nestjs/common";
 
+/**
+ * @brief This is where all the lobby requests are handled
+ */
 @WebSocketGateway()
 export class LobbyGateway {
   @WebSocketServer()
@@ -20,6 +24,14 @@ export class LobbyGateway {
 
   constructor(protected readonly lobbyService: LobbyService) {}
 
+  /**
+   * @brief Listen to CreateLobby event and proceed to creation when triggered
+   * @param client The client who requested the lobby creation
+   * @param body request body, protected by a validation pipe
+   * (see {@link LobbyValidationPipe} and {@link LobbyDto} for more information
+   * @return Dispatch the lobby creation result status to the client
+   * with the lobbyId if it appliesk
+   */
   @SubscribeMessage(ClientEvents.CreateLobby)
   onCreateLobby(
     @ConnectedSocket() client: AuthenticatedSocket,
@@ -46,7 +58,7 @@ export class LobbyGateway {
   @SubscribeMessage(ClientEvents.JoinLobby)
   onJoinLobby(
     client: AuthenticatedSocket,
-    data: any
+    @MessageBody(new ValidationPipe()) data: JoinLobbyDto
   ): WsResponse<ServerPayloads[ServerEvents.LobbyMessage]> {
     this.lobbyService.join(data.lobbyId, client);
 
@@ -62,42 +74,22 @@ export class LobbyGateway {
     };
   }
 
-  // @SubscribeMessage(ClientEvents.InviteToLobby)
-  // onInviteToLobby(
-  //   client: AuthenticatedSocket,
-  //   data: LobbyInviteDto
-  // ): WsResponse<ServerPayloads[ServerEvents.LobbyMessage]> {
-  //   const lobby = client.data.gameLobby;
-  //   if (!lobby) throw new Error(`No lobby attached to user - [${client.data.name}]`);
-  //   else lobby.inviteClient(client, data.invitedClient);
-  //   return {
-  //     event: ServerEvents.LobbyMessage,
-  //     data: {
-  //       message: `Invitation sent`,
-  //     },
-  //   };
-  // }
+  @SubscribeMessage(ClientEvents.LeaveLobby)
+  onLeaveLobby(
+    client: AuthenticatedSocket,
+    @MessageBody(new ValidationPipe()) data: LeaveLobbyDto
+  ): WsResponse<ServerPayloads[ServerEvents.LobbyMessage]> {
+    this.lobbyService.leave(data.lobbyId, client);
 
-  // @SubscribeMessage(ClientEvents.InvitationResponse)
-  // onInvitationResponse(
-  //   client: AuthenticatedSocket,
-  //   data: any
-  // ): WsResponse<ServerPayloads[ServerEvents.LobbyMessage]> {
-  //   console.log(data);
-  //   switch (data.status) {
-  //     case "accepted":
-  //       this.lobbyService.join(data.lobbyId, client);
-  //       break;
-  //     case "declined":
-  //       this.lobbyService.delete(data.lobbyId);
-  //   }
-  //   this.lobbyService.dispatchResponse(data.lobbyId, data.status);
-  //   return {
-  //     event: ServerEvents.InvitationResponse,
-  //     data: {
-  //       message: `Invitation ${data.status}`,
-  //       status: data.status,
-  //     },
-  //   };
-  // }
+    console.info(
+      `User - [${client.data.name}] - left the lobby - [${data.lobbyId}]`
+    );
+
+    return {
+      event: ServerEvents.LobbyMessage,
+      data: {
+        message: "Lobby left",
+      },
+    };
+  }
 }
