+++ backend/src/lobby/lobby.service.ts
@@ -1,12 +1,20 @@
 import { Inject, Injectable } from "@nestjs/common";
-import { AuthenticatedSocket } from "./types/server.type";
+import { AuthenticatedSocket } from "./types/lobby.type";
 import { ALobby } from "./ALobby";
 import { Server } from "socket.io";
 import { WebSocketServer } from "@nestjs/websockets";
-import {TLobbyDto} from "./dtos/lobby.dto";
+import {TLobbyDto} from "./dto/lobby.dto";
 
+/**
+ * @brief This service manage all the lobby instances on the server
+ * @param lobbyCreator The factory used to create different lobby types
+ * @member lobbies The list of lobbies accessible via their ids
+ */
 @Injectable()
 export class LobbyService {
+
+  constructor(@Inject("LOBBY_FACTORY") private readonly lobbyCreator: any) {}
+
   private readonly lobbies: Map<ALobby["id"], ALobby> = new Map<
     ALobby["id"],
     ALobby
@@ -14,8 +22,6 @@ export class LobbyService {
 
   @WebSocketServer()
   server: Server;
-
-  constructor(@Inject("LOBBY_FACTORY") private readonly lobbyCreator: any) {}
 
   public getLobby(lobbyId: string): ALobby {
     console.log(`Get lobbies = `, this.lobbies.keys());
